backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
scikit-learn
joblib
numpy
